Split Item implementation into AbstractItem and Item.
